+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Adminer dev
 - Autocomplete SQL commands
 - Do not edit NULL values by Modify (bug #967)
+- Fix foreign key actions (regression from 5.1.1)
 - PostgreSQL: Support COPY FROM stdin in SQL query (bug #942)
 - MySQL: Display number of found rows in group queries (regression from 5.1.1)
 - non-MySQL: Parse '--' without trailing space as comment in SQL command (bug SF-842)